# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Classical Music Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: Browse all listings
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [lesson, event, masterclass, digital_product]
        - name: instrument
          in: query
          schema:
            type: string
        - name: skillLevel
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: integer
        - name: maxPrice
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingListResponse"

    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      security: [{ clerkAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingBody"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /listings/teacher/{userId}:
    get:
      operationId: getTeacherListings
      tags: [listings]
      summary: Get all listings for a teacher
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Teacher's listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingListResponse"

  # ── Masterclasses ──────────────────────────────────────────────────────────
...
  /digital-products:
    get:
      operationId: listDigitalProducts
      tags: [digital-products]
      summary: Browse digital products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: instrument
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of digital products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalProductListResponse"

    post:
      operationId: createDigitalProduct
      tags: [digital-products]
      summary: Create a digital product listing
      security: [{ clerkAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDigitalProductBody"
      responses:
        "201":
          description: Created digital product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalProduct"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /digital-products/{id}:
    get:
      operationId: getDigitalProduct
      tags: [digital-products]
      summary: Get a digital product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Digital product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalProduct"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateDigitalProduct
      tags: [digital-products]
      summary: Update a digital product
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDigitalProductBody"
      responses:
        "200":
          description: Updated digital product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalProduct"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Bookings ───────────────────────────────────────────────────────────────
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for current user
      security: [{ clerkAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking request
      security: [{ clerkAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Orders ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── User ────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
          enum: [teacher, student, null]
        createdAt:
          type: string
          format: date-time
      required: [id, email, createdAt]
...
    TeacherProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        bio:
          type: ["string", "null"]
        instruments:
          type: array
          items:
            type: string
        genres:
          type: array
          items:
            type: string
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        hourlyRate:
          type: ["integer", "null"]
        currency:
          type: string
        yearsExperience:
          type: ["integer", "null"]
        education:
          type: ["string", "null"]
        averageRating:
          type: integer
        reviewCount:
          type: integer
        isVerified:
          type: boolean
        profileImageUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        videoIntroUrl:
          type: ["string", "null"]
        stripeOnboarded:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, instruments, genres, currency, averageRating, reviewCount, isVerified, stripeOnboarded, createdAt]
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        teacherId:
          type: string
        type:
          type: string
          enum: [lesson, event, masterclass, digital_product]
        status:
          type: string
          enum: [active, inactive, draft]
        title:
          type: string
        description:
          type: ["string", "null"]
        instrument:
          type: ["string", "null"]
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced, all]
        priceInCents:
          type: integer
        currency:
          type: string
        durationMinutes:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        isOnline:
          type: boolean
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        teacher:
          $ref: "#/components/schemas/TeacherProfile"
        createdAt:
          type: string
          format: date-time
      required: [id, teacherId, type, status, title, skillLevel, priceInCents, currency, tags, isOnline, createdAt]
    ListingListResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
      required: [listings, total]
    CreateListingBody:
      type: object
      properties:
        type:
          type: string
          enum: [lesson, event, masterclass, digital_product]
        title:
          type: string
        description:
          type: string
        instrument:
          type: string
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced, all]
        priceInCents:
          type: integer
        durationMinutes:
          type: integer
        imageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        isOnline:
          type: boolean
        city:
          type: string
        country:
          type: string
      required: [type, title, priceInCents]
    UpdateListingBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, inactive, draft]
        title:
          type: string
        description:
          type: string
        instrument:
          type: string
        skillLevel:
          type: string
        priceInCents:
          type: integer
        durationMinutes:
          type: integer
        imageUrl:
          type: string
        tags:
          type: array
          items:
            type: string

    # ── Masterclass ───────────────────────────────────────────────────────────
...
    DigitalProduct:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        teacherId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        instrument:
          type: ["string", "null"]
        difficulty:
          type: ["string", "null"]
        fileKey:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        fileType:
          type: ["string", "null"]
        previewUrl:
          type: ["string", "null"]
        downloadCount:
          type: integer
        isPublished:
          type: boolean
        priceInCents:
          type: integer
        currency:
          type: string
        teacher:
          $ref: "#/components/schemas/TeacherProfile"
        createdAt:
          type: string
          format: date-time
      required: [id, listingId, teacherId, title, category, downloadCount, isPublished, priceInCents, currency, createdAt]
    DigitalProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/DigitalProduct"
        total:
          type: integer
      required: [products, total]
    CreateDigitalProductBody:
      type: object
      properties:
        listingId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        instrument:
          type: string
        difficulty:
          type: string
        priceInCents:
          type: integer
        previewUrl:
          type: string
      required: [listingId, title, category, priceInCents]
    UpdateDigitalProductBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        instrument:
          type: string
        difficulty:
          type: string
        priceInCents:
          type: integer
        previewUrl:
          type: string
        isPublished:
          type: boolean

    # ── Booking ───────────────────────────────────────────────────────────────
    Booking:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: string
        teacherId:
          type: string
        listingId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [lesson, event]
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed, refunded]
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        durationMinutes:
          type: ["integer", "null"]
        priceInCents:
          type: integer
        currency:
          type: string
        notes:
          type: ["string", "null"]
        meetingUrl:
          type: ["string", "null"]
        instrument:
          type: ["string", "null"]
        eventType:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]
          format: date-time
        eventLocation:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        teacher:
          $ref: "#/components/schemas/TeacherProfile"
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, teacherId, type, status, priceInCents, currency, createdAt]
    BookingListResponse:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
      required: [bookings, total]
    CreateBookingBody:
      type: object
      properties:
        teacherId:
          type: string
        listingId:
          type: integer
        type:
          type: string
          enum: [lesson, event]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        notes:
          type: string
        instrument:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
          format: date-time
        eventLocation:
          type: string
      required: [teacherId, type]
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, cancelled, completed]
        meetingUrl:
          type: string
        cancelReason:
          type: string
        scheduledAt:
          type: string
          format: date-time

    # ── Order ─────────────────────────────────────────────────────────────────
...